% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EvoTraceR_object}
\alias{EvoTraceR_object}
\title{example data obtained from running function \code{asv_analys}, where the original barcode is identified and all ASVs are aligned to it.}
\format{
EvoTraceR object updated with the following fields: 
\describe{
}
}
\usage{
data(EvoTraceR_object)
}
\description{
example data obtained from running function \code{asv_analys}, where the original barcode is identified and all ASVs are aligned to it.
}
