% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{infer_phylogeny}
\alias{infer_phylogeny}
\title{infer_phylogeny}
\usage{
infer_phylogeny(EvoTraceR_object, mutations_use = "del_ins")
}
\arguments{
\item{EvoTraceR_object}{(Required)}

\item{mutations_use}{(optional). Default = 'del_ins' A string indicating what type of mutations to use for phylogeny reconstruction.
Can be one of c('del', 'del_ins'). The first
corresponds to using only cleaned deletions. The second refers to the case where cleaned deletions
and insertions are considered.}
}
\description{
This function reconstructs the phylogenetic tree using the \href{https://github.com/YosefLab/Cassiopeia}{Cassiopeia} suite. 
It uses the greedy algorithm, that iteratively splits the set of ASVs in clusters
based on the most common mutation at each iteration. Note that this requires installing cassiopeia in the anaconda/virtual enviroment used by reticulate.
}
\examples{
\dontrun{
EvoTraceR_object = infer_phylogeny(EvoTraceR_object)
}

}
