% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{initialize_EvoTraceR}
\alias{initialize_EvoTraceR}
\title{initialize_EvoTraceR}
\usage{
initialize_EvoTraceR(
  output_dir,
  input_dir = NULL,
  dada2_output_sequences = NULL,
  output_dir_dada2 = NULL,
  random_seed = NULL,
  output_figures = TRUE,
  multithread = TRUE,
  map_file_sample = NULL,
  dada2_pooled_analysis = FALSE,
  dada2_chimeras_minFoldParentOverAbundance = 8,
  verbose = TRUE,
  dada2_errorRate_estimation = "default",
  ...
)
}
\arguments{
\item{output_dir}{(Required). Path to the directory where all output files will be stored. The following \code{.csv} files will be created:
\itemize{
\item \code{quality_track_reads.csv}: track of the number of sequences during the 
different steps of \code{dada2} analysis.
\item \code{ dada2_asv_prefilter.csv}: sequences detected by dada2, with the counts detected in each sample.
}}

\item{input_dir}{Path to the directory containing \code{.fastq} files for forward and reverse reads.
This folder should contain the fastq files (2 for each sample) with the following name pattern:
FILEPREFIX_SAMPLE_BARCODEVERSION_R1.fastq FILEPREFIX_SAMPLE_BARCODEVERSION_R1.fastq. SAMPLE refers to either an organ (in case multiple organs were sequenced)
or timepoint (if longitudinal data are provided). Note that EvoTraceR does not support mixed sample types (i.e. samples must be either all from organs or all from timepoints).
Required when \code{dada2_output_sequences} is null (i.e. when the user has not previously run dada2).}

\item{dada2_output_sequences}{(Optional). In case users have already run dada2 up to the removal of bimeras, they should provide the path to the csv file containing the output (See description).
The output should contain samples on rows and sequences on columns.}

\item{output_dir_dada2}{(Optional). Output folder for dada2 filtered fastqs. If it doesn't exist it is created. Default is NULL and results in creating a 
sub-folder in \code{output_folder} named 'filtered_fastq/'.}

\item{random_seed}{(Optional). Seed for reproducibility. Deafult is NULL.}

\item{output_figures}{(Optional). Boolean indicating whether or not to save intermediate dada2 plots such as read quality profiles.
Deafult is \code{TRUE} and corresponds to creating a sub-folder inside \code{output_dir} named "dada2_figures".
This folder will contain the following plots:
\itemize{
\item Quality profiles of both forward and reverse reads (generated by \code{dada2}).
\item Error rates (generated by \code{dada2}).
\item Histogram of sequences length.
}}

\item{multithread}{(Optional) default \code{TRUE}. Whether to enable multithreading. If TRUE the number of threads is determined automatically (dada2 functionality).
If an integer is given, the number of threads is determined by its value.}

\item{map_file_sample}{(Optional). In case fastq files names are not in the format \cr FILEPREFIX_SAMPLE_FILESUFFIX_RX.fastq),
then users should provide a list that associates each filename (without the suffix _R1 and _R2) to the corresponding organ/day.
(e.g., if we have forward and reverse files named file1_R1.fastq and file1_R2.fastq that correspond to organ PRL (code for Prostate Left), than the parameter should be set as: \code{map_file_sample = c("file1" = "PRL")}).}

\item{dada2_pooled_analysis}{(Optional). Deafault = FALSE. Boolean that is passed to dada2 function \code{dada}, which performs sample inference.
It can be set to TRUE in case of multiple samples coming from the same mouse (e.g. different organs or multiple time points) (See \href{https://benjjneb.github.io/dada2/pool.html}{here} for more information).}

\item{dada2_chimeras_minFoldParentOverAbundance}{(Optional). Deafult = 8. parameter passed to the \code{dada2} function \code{isBimeraDenovo} during the call to \code{dada2 removeBimeraDenovo}.}

\item{verbose}{(Optional). Default TRUE. Boolean indicating whether or not to print the text output of dada2 functions.}

\item{dada2_errorRate_estimation}{(Optional). Must be one of c('default', 'loessErrfun_mod1', 'loessErrfun_mod2', 'loessErrfun_mod3', 'loessErrfun_mod4').
In case the fastq quality scores are in the range 0-40, keep the default option. The other 4 functions need to be used when dealing with NovaSeq data,
where the quality scores in the fastq files are binned to only 4 different values. See \href{https://github.com/ErnakovichLab/dada2_ernakovichlab#learn-the-error-rates}{here} for more details.}

\item{...}{(Optional) Any additional parameters passed to \code{dada2} functions.}
}
\value{
An object of type EvoTraceR, which is a list that will contain the following fields: 
\itemize{
\item \code{fastq_directory}: directory where the input fastq files are located.
\item \code{output_directory}: directory where all the output files are being stored.
\item \code{map_file_sample}: dataframe has as many rows as the input datasets, and for each input stores the sample (e.g. organ or day for longitudinal data)
to which it is associated.
\item \code{dada2_asv_prefilter}: dataframe that stores all sequences detected by \code{dada2}. Note that
these sequences still need to be filtered (see also \code{\link{asv_analysis}}).
\item \code{dada2}: list which contains the percentage of chimeras found by \code{dada2} and a dataframe
that tracks the number of sequences during all \code{dada2} steps.
}. This function also saves the following \code{.csv files} in the subfolder \code{dada2_files} created in the output dirctory provided by the user:
\itemize{
\item dada2_asv_prefilter.csv: sequences detected by dada2, with the counts detected in each sample.
\item quality_track_reads.csv: track of the number of sequences during all \code{dada2} steps.
}
}
\description{
This function initializes the EvoTraceR object, by computing the set of Amplicon Sequence Variants.
It wraps all the steps performed by dada2, that are the following:
(1) Filter and trimming reads.
(2) Learning the error rates.
(3) Dereplication.
(4) Sample Inference.
(5) Merging forward and reverse reads.
(6) Construction of the Amplicon Sequence Variant table (ASV). 
(7) Removal of chimeras.
(8) [Optional] Production of a summary which tracks changes in the number of reads at every step of the pipeline.
}
\details{
For further information please refer to \href{https://benjjneb.github.io/dada2/tutorial_1_8.html}{the original vignette}.
If users wish to manually run each step of dada2 computation they should execute all steps reported
in the vignette, up to the Removal of chimeras and save the output of \code{removeBimeraDenovo} in a csv file and provide the path to that file as parameter \code{dada2_output_sequences}.
}
\examples{
input_dir = system.file("extdata", "input", package = "EvoTraceR")
output_dir = system.file("extdata", "output", package = "EvoTraceR")
initialize_EvoTraceR(input_dir = input_dir, output_dir = output_dir)

}
