% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exported.R
\name{plot_cluster_summary}
\alias{plot_cluster_summary}
\title{plot_cluster_summary}
\usage{
plot_cluster_summary(EvoTraceR_object, cluster_list = NULL)
}
\arguments{
\item{EvoTraceR_object}{(Required).}

\item{cluster_list}{(Optional). Subset of clusters for which one wants to create the visualization.}
}
\value{
NULL. This function stores in the phylogeny output directory a pdf file for each cluster.
}
\description{
This function plots the phylogenetic tree and the frequency bubble plot for each cluster identified in the phylogenetic tree.
}
\examples{
\dontrun{
data(revo_phyl)
output_dir = system.file("extdata", "output", package = "EvoTraceR")
revo_phyl$output_directory = output_dir
summary_plot = plot_cluster_summary(revo_phyl)
}

}
