% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revo_analyzed}
\alias{revo_analyzed}
\title{example data obtained from running function \code{asv_analys}, where the original barcode is identified and all remaining ASVs are aligned to it.}
\format{
EvoTraceR object updated with the following fields: 
#' \describe{
\item{clean_asv_dataframe}{ASV sequences identified post-filtering (contamination removed,
sequences with a similarity higher than \code{pid_cutoff_nmbc} to the original barcode
aggregated to it and ASVs named in increasing order (ASV01, ASV02, etc.) according
to their total counts. }
\item{barcode}{Info about the barcode selected for the current analysis.}

 \item{statistics}{another list with the following sub-fileds: 
\describe{

\item{asv_df_percentages}{dataframe with six columns. \code{asv_names} is the name of the ASV.
\code{day_organ} is the sample identifier (e.g. ID of an organ or, in case of longitudinal data, of the timepoint);
\code{count}: total counts for a specific ASV in a specific sample;
\code{perc_in_sample}: counts normalized to the total counts in the corresponding sample;
\code{perc_asv}: conuts normalized to the total counts for the corresponding ASV;
\code{perc_fold_to_max}: counts normalized to the maximum counts observed for the corresponding ASV in a sample.}

\item{asv_totalCounts}{for each ASV, total counts and number of samples in which it was detected.}
\item{sample_totalCounts}{for each sample, total counts and number of distinct ASVs detected.}
\item{asv_diversity_perSample}{measures of clonal richness and measures of heterogeneity computed for each sample based on the ASVs detected.}
\item{asv_persample_frequency}{counts for each ASV in each sample.}
\item{asv_persample_detection}{binary matrix indicating whether a sequence has been detected in the corresponding sample.}
\item{asv_toBarcode_similarity}{edit distance, percentage similarity and alignment score of each ASV compared to the original barcode.}
}}
}
}
\usage{
data(revo_analyzed)
}
\value{
EvoTraceR object updated with statistics about the ASVs.
}
\description{
example data obtained from running function \code{asv_analys}, where the original barcode is identified and all remaining ASVs are aligned to it.
}
