% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revo_msa}
\alias{revo_msa}
\title{example data obtained from running function \code{compute_msa}, where alterations are identified in all the ASvs.}
\format{
EvoTraceR object with a new field named \code{alignment}, which is a list with the following fields:
\describe{
\item{msa_stringset}{output of MSA performed with MUSCLE}
\item{mutations_df}{tibble where each line corresponds to a position in a ASV, and the columns encode the name of the ASV, the sample
the position of the alteration in the original barcode, the type of alteration and the percentages of sequences that map to a sample and display the alteration.
}
}
}
\usage{
data(revo_msa)
}
\description{
example data obtained from running function \code{compute_msa}, where alterations are identified in all the ASvs.
}
