% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{setup_reticulate}
\alias{setup_reticulate}
\title{setup_reticulate}
\usage{
setup_reticulate()
}
\description{
This function loads reticulate and needs to be run after the installation of the package
}
