% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{analyse_mutations}
\alias{analyse_mutations}
\title{analyse_mutations}
\usage{
analyse_mutations(EvoTraceR_object)
}
\arguments{
\item{EvoTraceR_object}{EvoTraceR object on which we want to perform msa.}
}
\value{
EvoTraceR object with the field \code{alignment}, updated with 
\item \code{mutations_df}: dataframe containing for each position in each ASV the corresponding mutation state and the frequency of that mutation in all sequences.
}
\description{
This function considers the pairwise alignment output performed with \code{asv_analysis} and computes the overall frequency in each 
sample of each mutation in each barcode position.
}
\examples{
data(revo_analyzed)
output_dir = system.file("extdata", "output", package = "EvoTraceR")
revo_analyzed$output_directory = output_dir
revo_msa = analyse_mutations(revo_analyzed)

}
