% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{analyse_mutations}
\alias{analyse_mutations}
\title{analyse_mutations}
\usage{
analyse_mutations(EvoTraceR_object)
}
\arguments{
\item{EvoTraceR_object}{EvoTraceR object.}
}
\value{
EvoTraceR object with the field \code{alignment}, updated with 
\code{mutations_df}: dataframe containing for each position in each ASV the corresponding mutation state and the frequency of that mutation in all sequences.
}
\description{
This function considers the pairwise alignment output performed with \code{asv_analysis} and computes the overall frequency in each 
sample of each mutation in every barcode position.
}
\examples{
\dontrun{
data(EvoTraceR_analyzed)
EvoTraceR_object = analyse_mutations(EvoTraceR_object)
}

}
