% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{asv_analysis}
\alias{asv_analysis}
\title{ASV_analysis}
\usage{
asv_analysis(
  EvoTraceR_object,
  ref_name = "BC10v0",
  ref_seq =
    "TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTTTTACACACGCATTCAACCACGGACTGCTACACACGCACTCAACCGTGGATATTTACATACTCGTTCAACCGTGGATTGTTACACCCGCGTTCAACCAGGGTCAGATACACCCACGTTCAACCGTGGTACTATACTCGGGCATTCAACCGCGGCTTTCTGCACACGCCTACAACCGCGGAACTATACACGTGCATTCACCCGTGGATC",
  ref_flank_left = "^TCTAC",
  ref_flank_right = "CCCGTGGATC$",
  flanking_filtering = "right",
  ref_cut_sites = c(17, 42, 68, 94, 120, 146, 171, 198, 224, 251),
  ref_border_sites = c(1, 26, 52, 78, 104, 130, 156, 182, 208, 234),
  output_figures = TRUE,
  asv_count_cutoff = 2,
  pwa_gapOpening = -25,
  pwa_gapExtension = 0,
  pwa_match = 15,
  pwa_mismatch = -4,
  pwa_type = "global",
  cleaning_window = c(3, 3)
)
}
\arguments{
\item{EvoTraceR_object}{(Required). Object of class EvoTraceR, result of the function \code{initialize_EvoTraceR}}

\item{ref_name}{String indicating the ID of the reference sequence used in the experiment. Default is 'BC10v0',}

\item{ref_seq}{String indicating the reference sequence used in the experimenti. Default is 'TCTACACGCGCGTTCAACCGAGGAAAACTACACACACGTTCAACCACGGTTTTTTACACACGCATTCAACCACGGACTGCTACACACGCACTCAACCGTGGATATTTACATACTCGTTCAACCGTGGATTGTTACACCCGCGTTCAACCAGGGTCAGATACACCCACGTTCAACCGTGGTACTATACTCGGGCATTCAACCGCGGCTTTCTGCACACGCCTACAACCGCGGAACTATACACGTGCATTCACCCGTGGATC',}

\item{ref_flank_left}{String indicating the first nucleotides of the reference sequence that never mutate over the
course of the experiment. Default is "^TCTAC",}

\item{ref_flank_right}{String indicating the first nucleotides of the reference sequence that never mutate over the
course of the experiment. Default is "CCCGTGGATC$",}

\item{flanking_filtering}{Which among the flaning regions to use to filter out contaminated sequences.
Must be one of c('left', 'right', 'both', 'either'), default is 'both'}

\item{ref_cut_sites}{Positions in the reference sequence of the cutting sites. Default is c(17, 42, 68, 94, 120, 146, 171, 198, 224, 251),}

\item{ref_border_sites}{c(26, 52, 78, 104, 130, 156, 182, 208, 234).}

\item{output_figures}{(Optional). Deafult TRUE: Boolean indicating whether a user whishes to store a figure indicating the number of ASV tracked during the different steps of the analysis.}

\item{asv_count_cutoff}{(Optional). Default to 2. Minimum number of counts for an ASV to be considered in the statistics.}

\item{pwa_gapOpening}{(Optional). Delafult is -25. Parameter \code{gapOpening} passed to \code{pairwiseAlignment} from \code{Biostrings} (See description).}

\item{pwa_gapExtension}{(Optional). Default is 0. Parameter \code{gapExtension} passed to \code{pairwiseAlignment} from \code{Biostrings} (See description). Default is 0.}

\item{pwa_match}{(Optional). Default is 15. Parameter indicating the score for matches during pairwise alignment with \code{Biostrings}. This parameter,
together with the previous one, are used to construct the substitution matrix used by the function \code{pairwiseAlignment}.}

\item{pwa_mismatch}{(Optional). Default is -4. Parameter indicating the penalty for mismatch events during pairwise alignment with \code{Biostrings}.}

\item{pwa_type}{(Optional). Parameter indicating the type of pairwise alignment. Must be one of One of "global", "local", "overlap", "global-local", and "local-global".
For more details see \href{https://www.rdocumentation.org/packages/Biostrings/versions/2.40.2/topics/pairwiseAlignment}{original documentation}}

\item{cleaning_window}{(Optional). Deafult is c(3,3). Vector containing the number of nucleotides that we use for extending respectively the start and end position of each indel to determine the ones that don't span any cut sites and thus get removed. (See description for more information).}
}
\value{
The EvoTraceR object passed as a parameter with the following new fields:
\itemize{
\item \code{clean_asv_dataframe}: ASV sequences identified post-filtering (contamination removed,
sequences with a similarity higher than \code{pid_cutoff_nmbc} to the original barcode
aggregated to it and ASVs named in increasing order (ASV01, ASV02, etc.) according
to their total counts. 
\item \code{reference}: Info about the reference sequence used for the current analysis.

 \item \code{statistics}: another list with the following sub-fields: 
\itemize{

\item \code{asv_df_percentages}: dataframe with six columns. \code{asv_names} is the name of the ASV.
\code{sample} is the sample identifier (e.g. ID of an organ or, in case of longitudinal data, of the timepoint);
\code{count}: total counts per million for a specific ASV in a specific sample;
\code{perc_in_sample}: CPM normalized to the total counts in the corresponding sample;
\code{perc_asv}: CPM normalized to the total counts for the corresponding ASV;
\code{perc_fold_to_max}: CPM normalized to the maximum counts observed for the corresponding ASV in a sample.

\item \code{asv_totalCounts}: for each ASV, total counts and number of samples in which it was detected.
\item \code{sample_totalcounts}: for each sample, total counts and number of distinct ASVs detected.
\item \code{asv_diversity_persample}: measures of clonal richness and measures of heterogeneity computed for each sample based on the ASVs detected.
\item \code{asv_persample_frequency}: counts for each ASV in each sample.
\item \code{asv_persample_detection}: binary matrix indicating whether a sequence has been detected in the corresponding sample.
\item \code{asv_toBarcode_similarity}: edit distance, percentage similarity and alignment score of each ASV compared to the original barcode.
\item \code{all_asv_statistics}: all the statistics computed on each ASV grouped together in the same tibble.
}
\item \code{alignment}, another list with the following fields:
\itemize{
\item \code{Binary_mutation_matrix}: binary matrix encoding the presence/absence of a mutation in an ASV.
\item \code{asv_barcode_alignment}: tibble where each line corresponds to a position in a ASV, and the columns encode the following information:
\itemize{
\item asv_names: name of the ASV
\item sample: sample identifier
\item position_bc260: position of the alteration in the original barcode. Note that insertions
are assigned to the position that coincides with their beginning.
\item alt: type of alteration. wt = Wild Type (i.e. non-mutated position). sub = substitution. del = deletion. ins = insertion.
\item{ref_asv}, \item{read_asv}: respectively, the reference nucleotide observed in the original barcode and the one observed on the sequence.
}
\item \code{mutations_coordinates}: dataframe containing a list of all mutations, with their start and end position.
\item \code{ASV_alterations_width}: dataframe containing the number of nucleotides affected by each type of mutation in each ASV.
}
}
This function saves the figure \code{asv_filtering_freq.pdf}, which is a barplot indicating the change in the number of sequences 
during all the processing steps.
}
\description{
This function performs the analysis on ASV sequences identified by the previous steps and aligns them to the reference sequence.
First, it pools together those sequences characterized by a Hamming distance equals or lower than 2, summing their counts. To perform this step, EvoTraceR uses the
sequence clusering agorithm impemented in the python package \href{https://umi-tools.readthedocs.io/en/latest/QUICK_START.html}{UMI-tools}
Then it performs pairwise alignment using Needleman-Wunsch global alignment algorithm implemented in function \code{pairwiseAlignment}
in package \code{Biostrings}, aligning each sequence to the original barcode considered in the analysis.
(See the Biostrings documentation \href{https://www.rdocumentation.org/packages/Biostrings/versions/2.40.2/topics/pairwiseAlignment}{here} for more details).
After identifying all indels (insertions are identified by their start position and number of nucleoides inserted, while deletions are identified by their start and end position),
it removes the ones that are too small and don't span any cut site: to perform this filter, it exapnds the start and end position by the number of bases specified in the parameter \code{cleaning_window},
it counts the number of cut sites spanned after the expantion and it removes those that span 0 sites.
Next, it pools together those sequences that differ by one another only by substitutions, summing their counts, and then it removes
those sequences showing a frequency lower than the threshold specified in parameter \code{asv_count_cutoff}.
Next, it performs filtering of the sequences based on their flanking sequences and finally it computes the normalized counts of each ASV in each sample,
dividing each count by the total number of sequences for each sample and multiplying by 1e6, yelding Counts Per Million (CPM).
}
\details{
Then, using the CPM it computes different statistics for the final ASVS, storing:
 the relative frequency of all ASVs in each sample. 
 the relative frequency of each ASV in the samples.
 the counts for each ASV normalized to the counts of the sample with maximum frequency.
 the frequency of the different ASVs in each sample.
}
\examples{
\dontrun{
data(EvoTraceR_object)
EvoTraceR_object = asv_analysis(EvoTraceR_object = EvoTraceR_object)
}

}
