% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{create_df_summary}
\alias{create_df_summary}
\title{create_df_summary}
\usage{
create_df_summary(EvoTraceR_object)
}
\arguments{
\item{EvoTraceR_object}{(Required).}
}
\value{
EvoTraceR_object with a field named plot_summary containing dataframe df_to_plot_final.
This dataframe containins all the information computed within the package for each ASV in each sample.
}
\description{
This function combines in one dataframe all the data that have been computed for the phylogenetic analysis.
}
\examples{
\dontrun{
EvoTraceR_object = create_df_summary(EvoTraceR_object)
}

}
