% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{infer_phylogeny}
\alias{infer_phylogeny}
\title{infer_phylogeny}
\usage{
infer_phylogeny(REvoBC_object, phylip_package_path)
}
\arguments{
\item{REvoBC_object}{(Required)}

\item{phylip_package_path}{(Required). Prior to running this function, users should install 
\href{https://evolution.genetics.washington.edu/phylip.html}{PHYLIP} and then provide the path to the folder containing the executable of mix.
For example, if in Windows a user stores the folder of phylip in path \code{D:/Programs/phylip},
then the value of \code{phylip_package_path} should be set to \code{D:/Programs/phylip/exe/}.}
}
\value{
REvoBC object with a new filed called phylogeny, that stores the inferred tree.
}
\description{
This function calls Rmix from package Rphylip to perform the inference of the phylogeny
with Camin-Sokal.
}
\examples{
\dontrun{
data(revo_msa)
revo_phyl = infer_phylogeny(revo_msa, phylip_package_path = 'C:/Users/*/Program Files/rphylip/exe/')
}


}
