% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{infer_phylogeny}
\alias{infer_phylogeny}
\title{infer_phylogeny}
\usage{
infer_phylogeny(EvoTraceR_object, phylip_package_path, mutations_use = "del_ins")
}
\arguments{
\item{EvoTraceR_object}{(Required)}

\item{phylip_package_path}{(Required). Prior to running this function, users should install 
\href{https://evolution.genetics.washington.edu/phylip.html}{PHYLIP} and then provide the path to the folder containing the executable of mix.
For example, if in Windows a user stores the folder of phylip in path \code{D:/Programs/phylip},
then the value of \code{phylip_package_path} should be set to \code{D:/Programs/phylip/exe/}.}

\item{mutations_use}{(optional). Default = 'smooth_del' A string indicating what type of mutations to use for phylogeny reconstruction.
Can be one of c('del', 'del_ins'). The first
corresponds to using only cleaned deletions. The second refers to the case where cleaned deletions
and insertios are considered.}
}
\value{
EvoTraceR object with a new filed called phylogeny, that stores the inferred tree.
}
\description{
This function calls Rmix from package Rphylip to perform the inference of the phylogeny
with Camin-Sokal.
}
\examples{
\dontrun{
data(revo_msa)
output_dir = system.file("extdata", "output", package = "EvoTraceR")
revo_msa$output_directory = output_dir
revo_phyl = infer_phylogeny(revo_msa, phylip_package_path = 'D:/Programmi/phylip-3.698/exe/')
}

}
