% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revo_initialized}
\alias{revo_initialized}
\title{example data obtained from running function \code{initialize_EvoTraceR}, where Trimmomatic and Flash are run on the fastqs provided in input.}
\format{
An object of class EvoTraceR, which is a list with the following fields:
\describe{
\item{fastq_directory}{directory where the input fastq files are located.}
\item{output_directory}{directory where all the output files are being stored.}
\item{map_file_sample}{dataframe has as many rows as the input datasets, and for each input stores the sample (e.g. organ or day for longitudinal data)
to which it is associated.}
\item{asv_prefilter}{dataframe that stores all sequences detected after this preliminary steps. Note that
these sequences still need to be filtered (see also \code{\link{asv_analysis}}).}
}. 
An object of class EvoTraceR
}
\usage{
data(revo_initialized)
}
\description{
example data obtained from running function \code{initialize_EvoTraceR}, where Trimmomatic and Flash are run on the fastqs provided in input.
}
