% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exported.R
\name{seq_filtering_plot}
\alias{seq_filtering_plot}
\title{plot_cluster_summary}
\usage{
seq_filtering_plot(
  EvoTraceR_object,
  figure_dir = EvoTraceR_object$output_directory
)
}
\arguments{
\item{EvoTraceR_object}{(Required).}

\item{what}{(Optional). Default to 'phylogeny'. Can be one of \code{c('phylogeny', 'frequency', 'pid', 'asv_length', 'alterations width', 'msa')}}
}
\value{
NULL. This function stores in the phylogeny output directory a pdf file for each cluster.
}
\description{
This function plots different aspects of the data.
}
\examples{
\dontrun{
data(revo_phyl)
output_dir = system.file("extdata", "output", package = "EvoTraceR")
revo_phyl$output_directory = output_dir
summary_plot = plot_cluster_summary(revo_phyl)
}

}
