% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exported.R
\name{seq_filtering_plot}
\alias{seq_filtering_plot}
\title{seq_filtering_plot}
\usage{
seq_filtering_plot(
  EvoTraceR_object,
  figure_dir = EvoTraceR_object$output_directory
)
}
\arguments{
\item{EvoTraceR_object}{(Required).}

\item{figure_dir}{(Optional). Directory where the output will be stored.}
}
\value{
NULL. This function stores in the phylogeny output directory a pdf file for each cluster.
}
\description{
This function plots the number of sequences left during all the preprocessing steps.
}
